@@ -7,9 +7,9 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.24.0",
-    "react": "^17.0.2",
-    "react-datepicker": "^4.5.0",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-datepicker": "^4.8.0",
+    "react-dom": "^18.1.0",
     "react-redux": "^7.2.6",
     "react-scripts": "^5.0.0",
     "react-switch": "^6.0.0",
